Analyzer/requirements.txt
paho-mqtt
influxdb-client
requests
pillow
opencv-python-headless